jwyang 신용카드할인수정/할인쿠폰사용시 DB입력수정

+++ hannol/src/main/java/com/sp/pay/PayServiceImpl.java
@@ -36,8 +36,12 @@ public class PayServiceImpl implements PayService {
 			dao.insertData("pay.insertPayInfo", dto);
 			dao.insertData("pay.insertCardInfo", dto);
 			dao.updateData("pay.insertPaymentInfo", dto);
-			dao.updateData("pay.insertGift", dto);
-			dao.deleteData("pay.deleteCart", dto);
+			if(dto.getParentCode() == 2) {
+				dao.updateData("pay.insertGift", dto);
+				dao.deleteData("pay.deleteCart", dto);
+			}else {
+				
+			}
 		} catch (Exception e) {
 			throw e;
 		}
